london_api2.py to catch the adress...
i transferred the part of the streamlit where you place an adress and an a current time to get to the firestations
which are the closest to the adress.
in the original there is a sb.csv, so i placed this into /data.

as a next step, i try to add the xgboostmodel to it.